Major improvements. Introduced context component, connection to services, guarder routes
Introduced context component, that is in charge of saving authentication state and providing it to other routes. The context component saves the user object, saves jwt and userId, that is later used for different purposes throughout the whole application.
Introduced guarded routes. The purpose of the guarder routes is to restrict access to parts of the application, when the user is not authenticated.  The guarder route utilises the outlet component and is passed as element with a child in the Route component.

The application is now connected some of the services (auth and projects service).