Finished the whole create project flow
Set up the different UI components for the create project flow. Also wrote the logic for contacting the server and sending the post request for creating the project.
Wrote logic for re-loading the page, when a new project is created, so the new project can be fetched immediately after creation. Additionally added a loader, when the DashBoard page component is fetching data.  This way the user doesn't wait on an empty screen.