Major updates
Finished the task creation flow, added a component, that represents a task, when it's expanded.
Moved elements into separate components.
Added tasks slider, that is the component, that wraps the task items and displays them as in a slider.
Added project slider, that is the component that wraps the project items and displays them as in a slider